Switches Counters to Handles

+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/Counter.java
@@ -65,7 +65,7 @@ public class Counter extends SensorBase implements CounterBase, LiveWindowSendab
   protected DigitalSource m_downSource; // /< What makes the counter count down.
   private boolean m_allocatedUpSource;
   private boolean m_allocatedDownSource;
-  private long m_counter; // /< The FPGA counter object.
+  private int m_counter; // /< The FPGA counter object.
   private int m_index; // /< The index of this counter.
   private PIDSourceType m_pidSource;
   private double m_distancePerPulse; // distance of travel for each tick
